# openapi.yaml
openapi: 3.1.0
info:
  title: PDF Accessibility Platform API
  description: |
    API for PDF accessibility processing platform

    This API provides document processing capabilities to convert PDFs into accessible formats
    including ADA/Section 508 compliant PDFs and LLM-ready structured data exports.

    ## Authentication
    - Bearer token authentication using BetterAuth JWT tokens
    - API keys for programmatic access (coming soon)

    ## Rate Limits
    - 60 requests per minute per user
    - 1000 requests per hour per user

    ## Supported Formats
    - Input: PDF, DOC, DOCX, TXT files
    - Output: Accessible PDF, HTML, JSON, CSV/ZIP exports

  version: 1.0.0
  contact:
    name: PDF Accessibility Platform
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.pdf-accessibility.com
    description: Production server
...
paths:
...
  /documents:
    post:
      tags: [documents]
      summary: Upload document for processing
      description: Upload a document file or provide URL for processing. Returns document ID and queues for processing.
      operationId: upload_document
      security:
        - BetterAuthJWTBearer: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Document file to upload
                source_url:
                  type: string
                  format: uri
                  description: URL to download document from
                filename:
                  type: string
                  description: Original filename
                priority:
                  type: boolean
                  default: false
                  description: High priority processing
                webhook_url:
                  type: string
                  format: uri
                  description: Callback webhook URL
                metadata:
                  type: string
                  description: JSON metadata string
      responses:
        '201':
          description: Document uploaded and queued for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'

    get:
      tags: [documents]
      summary: List user documents
      description: List documents for the authenticated user with pagination and optional status filtering
      operationId: list_documents
      security:
        - BetterAuthJWTBearer: []
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: status
          in: query
          description: Filter by document status
          schema:
            $ref: '#/components/schemas/DocumentStatus'
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    BetterAuthJWTBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from BetterAuth authentication
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
  schemas:
...
    DocumentStatus:
      type: string
      enum:
        [
          pending,
          processing,
          completed,
          failed,
          validation_failed,
          notification_failed,
        ]
      description: Document processing status
...
    DocumentResponse:
      type: object
      required: [doc_id, status, created_at, updated_at, user_id]
      properties:
        doc_id:
          type: string
          format: uuid
          description: Document ID
        status:
          $ref: '#/components/schemas/DocumentStatus'
        filename:
          type: string
          description: Original filename
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        completed_at:
          type: string
          format: date-time
          description: Completion timestamp
        user_id:
          type: string
          description: User who uploaded the document
        metadata:
          type: object
          additionalProperties: true
          description: Document metadata
        error_message:
          type: string
          description: Error message if failed
        artifacts:
          type: object
          additionalProperties:
            type: string
          description: Available artifacts
    DocumentListResponse:
      type: object
      required: [documents, total, page, per_page]
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentResponse'
          description: List of documents
        total:
          type: integer
          description: Total number of documents
        page:
          type: integer
          description: Current page number
        per_page:
          type: integer
          description: Items per page
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        details:
          type: object
          additionalProperties: true
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
          description: Error location
        msg:
          type: string
          description: Error message
        type:
          type: string
          description: Error type